SGRD-2 | add generate jwt token for public usage
